refactor(cli): extract onboarding command module (paths, mkdir, check, install, onboard)
install-plugin-assets.test.ts scanned cli/bakin.ts for the dispatch
source; it now scans the relocated onboarding command module (same
assertions, adjusted import-specifier path).
